Update public-networking.md custom domain section
Adding a line to Important Considerations to explain that Trial plan users only have one custom domain available. 
It currently isn't clear that www and root domains are considered unique in Railway. 
There are several posts in the Discussions and Discord asking about this so I'm trying to clarify directly in the docs.