fix faulty "Deploy a Laravel App" guide
guide suggested to call `php artisan optimize:clear` in `build-app.sh` that should used in "Custom Build Command". The thing is that Laravel have `CACHE_STORE=database` configured by default, it tries to access the DB, but `"postgres.railway.internal"` url is not available during Build phase, so build fails with a `QueryException`.