Update Rails deployment guide steps
Step 7 of a Deploy from Github repo broke my deployment. It set my application into an unknown state where I couldn't get any error messages or ssh into the machine. Very frustrating. The custom start command seems like a handy feature, but this tutorial is not the place for it.